+++ changelog.md
@@ -31,6 +31,7 @@ Codion Change Log
 - FilterColumnConditionPanel.Fields added.
 - FilterTableColumnModel.Visible and Hidden added, related refactoring.
 - FilterTableColumnModel.containsColumn() and resetColumns() renamed contains() and reset() respectively.
+- FilterColumnConditionPanel bug fixed, did not link all components to locked state.
 ### is.codion.framework.domain
 - Entity.copy(), copyBuilder() and deepCopy() moved to Entity.Copy, renamed mutable(), builder() and deep() respectively, immutable() added.
 - ImmutableEntity bug fixed, StackOverflowError when creating an immutable entity with cyclical foreign key dependencies, some refactoring.